requirements.txt
os
flask 
pillow
google-generativeai
